Markup change
@pysailor , I’ve changed the dfn tags (with hrefs??) into anchor tags.
Surely this was the intention? I’ve also aded the word ‘info’, as
otherwise the tag would be meaningless. The DFN scenario is only used
when there’s only content to be shown for which inline elements are
enough. The content of the tooltip is then the same as the content of
the DFN tag.