Improved model for warning text. Fixes #19
@pysailor, Allows for potentially multi line text. Not infinitely! Try
to keep it short :).

There’s an info icon that may link to the help section or represent a
helpt text.

I choose for a model that show the message on each screen, but it will
scroll away with the content so that it doesn’t use up too much space
on small screen all the time.

It has been tested on the screen ‘/preparation-test.html’ but must be
tested still on all other screen. This could more easily be done on the
dev version.