Let quaive-versions be a z3c.form. Start on quaive-publish form/view.
In Quaive when viewing a survey, you can now click on 'Publishing options'.
This injects @@quaive-versions.
There you can Cancel and inject brings you back the edit form.

Or you can click Publish and see @@quaive-publish.
This is a bit ugly for now.  And the Cancel and Publish buttons redirect you to OiRA.
So there is work there.

The Unpublish button points to a page that does not exist yet.