On @tool-versions: fix the reported review_state for surveys.
The review_state as reported by the workflow does not say everything currently.
A Survey (tool version) may have been published according to the workflow, but it may not be the version that is published on the client.
This should probably be fixed in Euphorie, but we can fix it here for the moment.