Add the @@survey-group-image view
This view is necessary to retrieve the image for the survey group in the
best possible way, given the constraints.

Survey groups are assigned via a script across all survey versions (see
https://github.com/syslabcom/scrum/issues/2552 for details).

The view redirects to the first meaningful image it can find. While this
approach helps achieve the desired functionality, it is more of a
workaround and isn't highly efficient. One downside is that browser
caching may retain the image longer than intended, potentially affecting
freshness.

This solution is a trade-off between performance and practicality.