Add quaive-publish page the text for publishing a survey to the client.
Nothing is actually done, this only displays text.  On the Quaive side, this text is shown with buttons to accompany it.

Originally on the `quaive-app` branch we had `@@quaive-versions` as well, but this was no longer called from Quaive.
This included a button pointing to `@@quaive-unpublish` as well, but this did not exist.
Anyway, both are not needed by Quaive.