Mailing Lists: Do catalog search in admin section
When first getting a catalog batch of client objects and then filtering
out some objects, we can end up with fewer items. This makes the
autosuggest think it is the last batch, and it stops requesting more.

We add a one state workflow to countries so that a query for `published`
also returns countries.

Ref https://github.com/syslabcom/scrum/issues/3124