OiRA Matomo: Add tracking snippet.
This tracking snippet deactivates automatic link tracking for the report
download buttons and tracks them manually.
It removes the session id segment from the URL so that all downloads can
be grouped together in Matomo and to not reveal DSGVO relevant data.

This requires Matomo 5+ as the deactivation does not work reliably
(Matomo 4) or at all (Matomo 3) in earlier Matomo versions.

Ref: scrum-3249