Add quaive-preview as wrapper around preview with a short template.
Alternatively, we could let Quaive call the standard `@@preview`, but we would need a way to select the few lines from the html that we actually need.
Selecting `#mainContent` mostly works, but it includes the title and buttons, and this needs updates to the css, because everything is too close together.