Do not use pat-inject in quaive XML export.
It interferes with the download: the attachment gets fetched by the Quaive side, and correctly passed on to the browser, but nothing shows up because pat-inject does not see the html ids that it is looking for.  So nothing happens and you see a spinner.

This is also why I went for a full form instead of a panel: I could not get it to work otherwise.