Miller columns: only show folders when they contain allowed content types anywhere.
This also means we need a different way to show the amount of items within an item for preview.  Previously we could simply count the results that we would show below the preview item anyway.  But now this list of results may be empty when none of the child objects of the preview item are selectable.  So we do an extra catalog query to get the number of child objects.

Second fix: show the items in the miller columns in the same order as in their folder.

Refs https://github.com/syslabcom/scrum/issues/4139